# module.js
const Category = require('./category');
const Product = require('./product');
const User = require('./user');
const Order = require('./order');

module.exports = {
  Category,
  Product,
  User,
  Order
};