# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "My Store API",
    "description": "API for my store",
    "version": "1.0.0"
  },
  "basePath": "/api",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/categories/{id}": {
      "get": {
        "summary": "Get category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Category ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A category object",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
...
    "/products/{id}": {
      "get": {
        "summary": "Get product by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "A product object",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number"
        },
        "category": {
          "$ref": "#/definitions/Category"
        }
      }
    },
...
  }
}